+++ src/main/java/com/talentsoft/recruitmentmoduleback/repository/CandidateRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CandidateRepository  extends JpaRepository<Candidate, Long> {
+public interface CandidateRepository  extends JpaRepository<Candidate, Integer> {
     @Query("SELECT c FROM Candidate c WHERE c.companyid = ?1 AND c.candidateStatusId != 7")
     List<Candidate> findAllByCompanyIdAndStatusNotEliminado(int companyId);
 }
